Deletes posts and their images from storage

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
@@ -127,11 +128,26 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
-      
+        //Handle file upload
+        if($request->hasFile('cover_image')){
+            //Get filename with ext.
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_image', $fileNameToStore);
+        }
         //create post
         $post =  Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
       
         return redirect('/posts')->with('success', 'Post Updated');
@@ -145,9 +161,17 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-        //
-        $post = Post::find($id);
+        
+        $post = Post::find($id);  
+        //Check for errors      
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        if($post->cover_image != 'noimage.jpg'){
+            //Delete image
+            Storage::delete('public/cover_image/'.$post->cover_image);
+        }
         $post->delete();
-        return redirect('/dashboard')->with('success', 'Post Deleted');
+        return redirect('/posts')->with('success', 'Post Removed');
     }
 }
